separate add terminal form
